better heuristic for modulename
Change ModulesForModules to take from,to. Generating a map of all the usages
before wasn't working because modules are stored with version (duh) so we can't
tell which version the package was referenced by. Package -> module only makes
sense in the context of from->to module edge where we know the version of the
module.

Also sorts list by ratio.